Imporved mailing system and added a trash can

+++ loggedIn/mails/index.php
@@ -4,41 +4,37 @@ require_once '../../webdev/php/Classes/ClassMail.php';
 require_once '../../webdev/php/Generators/tableGenerator.php';
 
 #Initing the objects
-$HTML = new HTMLGenertator\HTMLfile('Mail inbox', ['table.css', 'mails.css'], NULL, NULL, 1);
+$HTML = new HTMLGenertator\HTMLfile('Mail inbox', ['table.css', 'form.css', 'mails.css',], NULL, NULL, 1);
 $overview = new MailManager\Overview($_SESSION['studentId']);
 
+# Inting HTML file
 $HTML->outputHeader();
+
+# Initing mail vars
 $unreadCount = $overview->getUnread();
-switch($unreadCount){
-    case 0:
-	$insertString = 'no unread mails';
-	break;
-    case 1:
-	$insertString = '1 unread mail';
-	break;
-    default:
-	$insertString = $unreadCount . ' unread mails';
-}
+$idArray = $overview->getIds();
+$listedMails = count($idArray);
 ?>
-<a href="write.php">Create a new message</a>
-<h2>Inbox</h2>
-<h3>You have <?php echo $insertString; ?></h3>
-<table>
-    <?php
+<a href="write.php"><button>Create a new message</button></a>
+<h2>Inbox (<?php echo $unreadCount; ?>)</h2>
+<?php
+# Listing the mails
+if ($listedMails != 0) {
+    echo '<table id="mailTable">';
     echo generateTableHead(['Subject', 'Sender', 'X']);
-    $idArray = $overview->getIds();
-    for($i = 0; $i < count($idArray); $i++){
+    for ($i = 0; $i < $listedMails; $i++) {
 	$current = $overview->getMessage($idArray[$i]);
 	$id = $idArray[$i];
 	$unreadClass = ($current->isRead()) ? '' : ' class="unread"';
 	echo "<tr $unreadClass>";
 	echo '<td><a href="read.php?id=' . $id . '">' . $current->getSubject() . '</a></td>
-	<td>' . $current->getSender() . '</td>
-	<td><a href="delete.php?id=' . $id . '" class="delete">x</a></td>';
+	    <td>' . $current->getSender() . '</td>
+	    <td><a href="delete.php?id=' . $id . '" class="delete">x</a></td>';
 	echo '</tr>';
     }
-    ?>
-</table>
-<?php
+    echo'</table>';
+}else{
+    echo '<p>The inbox is empty!</p>';
+}
 $HTML->outputFooter();
 ?>
